# perf: optimize logging package