Scanning the codebase for some simple performance gains -- here's a 5x speedup of log level ID conversion:

---

1. Removes a map lookup in favor of a simple type switch.

```console
BenchmarkGetLevelID-12    	549921196	         2.106 ns/op	       0 B/op	       0 allocs/op
```

vs main
```console
BenchmarkGetLevelID-12    	100000000	        11.45 ns/op	       0 B/op	       0 allocs/op
```

---

2. Replaces a mutex+map with sync.Map

(Only a minor improvement in overall runs / about the same ns/ops)

```console
Benchmark_logOnce-12    	 8518489	       149.9 ns/op	     336 B/op	       2 allocs/op
```

vs main
```console
Benchmark_logOnce-12    	 7707423	       143.7 ns/op	     336 B/op	       2 allocs/op
```

---

3. Reduce allocations in logging package / optimize keypair sorting

```console
Benchmark_Info-12    	  581666	      2044 ns/op	     712 B/op	       9 allocs/op
```

vs main
```console
Benchmark_Info-12    	  490310	      2383 ns/op	     792 B/op	      14 allocs/op
```